#40 feat: 비밀번호 초기화 뷰 app 라우팅 전환
