+++ app/reset_password/[token]/page.tsx
@@ -1,0 +1,5 @@
+import ResetPasswordView from './ResetPasswordView';
+
+export default function ResetPassword() {
+    return <ResetPasswordView />;
+}
