+++ components/authentication/common/TextInput.tsx
@@ -1,3 +1,5 @@
+"use client";
+
 import { ChangeEvent, useState } from "react";
 
 export interface Props {
