--- pages_old/reset_password/[token]/index.tsx
@@ -1,5 +1,0 @@
-import ResetPasswordView from "components/reset_password";
-
-export default function ResetPasswordPage() {
-    return <ResetPasswordView />;
-}
